Rename LiveTvHost to RecordingsHost and move to recordings folder
